backend/requirements.txt
pypdf2
toml
langchain
transformers
torch
python-dotenv
pinecone-client
google-generativeai
fastapi[standard]
openai
langchain-openai
playwright
beautifulsoup4
langchain_community
bs4
html2text